+++ docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @naive-potato-ui/docs
 
+## 1.0.4
+
+### Patch Changes
+
+- change custom upload action
+
 ## 1.0.3
 
 ### Patch Changes